authentication

+++ src/controller/ProjectController.ts
@@ -8,7 +8,9 @@ export class ProjectController {
 
     const project = new Project(req.body);
 
- 
+    // Assign Manager to project
+    project.manager = req.user.id
+
     try {
       await project.save();
       return res.send('Projecto creado correctamente')
@@ -23,7 +25,11 @@ export class ProjectController {
 
     try {
 
-      const projects = await Project.find({})
+      const projects = await Project.find({
+        $or: [
+          { manager: { $in: req.user.id } }
+        ]
+      })
       res.json(projects)
 
     } catch (error) {
@@ -42,6 +48,11 @@ export class ProjectController {
 
       if (!project) {
         const error = new Error('No hay proyecto con ese ID')
+        return res.status(404).json({ error: error.message })
+      }
+
+      if (project.manager.toString() !== req.user.id.toString()) {
+        const error = new Error('Accion no valida')
         return res.status(404).json({ error: error.message })
       }
       res.json(project)
@@ -66,6 +77,12 @@ export class ProjectController {
       project.clientName = req.body.clientName
       project.projectName = req.body.projectName
       project.description = req.body.description
+
+      if (project.manager.toString() !== req.user.id.toString()) {
+        const error = new Error('Solo el manager puede eliminar el proyecto')
+        return res.status(404).json({ error: error.message })
+      }
+
       await project.save()
       res.send('Proyecto Actualizado')
 
@@ -86,6 +103,13 @@ export class ProjectController {
         const error = new Error('No hay proyecto con ese ID')
         return res.status(404).json({ error: error.message })
       }
+
+
+      if (project.manager.toString() !== req.user.id.toString()) {
+        const error = new Error('Accion no valida')
+        return res.status(404).json({ error: error.message })
+      }
+
       await project.deleteOne()
